+++ hbase-server/src/test/java/org/apache/hadoop/hbase/regionserver/TestRegionSplitPolicy.java
@@ -67,7 +67,7 @@ public class TestRegionSplitPolicy {
     // Configure IncreasingToUpperBoundRegionSplitPolicy as our split policy
     conf.set(HConstants.HBASE_REGION_SPLIT_POLICY_KEY,
       IncreasingToUpperBoundRegionSplitPolicy.class.getName());
-    conf.setDouble("hbase.hregion.max.filesize.jitter", 0.25D);
+    conf.setFloat("hbase.hregion.max.filesize.jitter", 0.25f);
 
     // Now make it so the mock region has a RegionServerService that will
     // return 'online regions'.
@@ -133,7 +133,7 @@ public class TestRegionSplitPolicy {
   @Test
   public void testCreateDefault() throws IOException {
     conf.setLong(HConstants.HREGION_MAX_FILESIZE, 1234L);
-    conf.setDouble("hbase.hregion.max.filesize.jitter", 0.25D);
+    conf.setFloat("hbase.hregion.max.filesize.jitter", 0.25f);
 
     // Using a default HTD, should pick up the file size from
     // configuration.
